split api and router
