+++ html/pfappserver/root/src/views/Auditing/adminApiLogs/_components/TheView.vue
@@ -25,11 +25,6 @@ export default {
     pfFormRow
   },
   props: {
-    storeName: { // from router
-      type: String,
-      default: null,
-      required: true
-    },
     id: String // from router
   },
   data () {
@@ -41,7 +36,7 @@ export default {
   },
   computed: {
     isLoading () {
-      return this.$store.getters[`${this.storeName}/isLoading`]
+      return this.$store.getters[`$_admin_api_audit_logs/isLoading`]
     },
     escapeKey () {
       return this.$store.getters['events/escapeKey']
@@ -49,7 +44,7 @@ export default {
   },
   methods: {
     init () {
-      this.$store.dispatch(`${this.storeName}/getItem`, this.id).then(item => {
+      this.$store.dispatch(`$_admin_api_audit_logs/getItem`, this.id).then(item => {
         this.item = item
       })
     },
