+++ html/pfappserver/root/src/views/Auditing/dhcpOption82Logs/_store.js
@@ -2,7 +2,7 @@
 * "$_dhcpoption82_logs" store module
 */
 import Vue from 'vue'
-import api from '../_api'
+import api from './_api'
 
 const STORAGE_SEARCH_LIMIT_KEY = 'dhcpoption82logs-search-limit'
 const STORAGE_VISIBLE_COLUMNS_KEY = 'dhcpoption82logs-visible-columns'
