+++ html/pfappserver/root/src/views/Auditing/dnsLogs/_store.js
@@ -2,7 +2,7 @@
 * "$_dns_logs" store module
 */
 import Vue from 'vue'
-import api from '../_api'
+import api from './_api'
 
 const STORAGE_SEARCH_LIMIT_KEY = 'dnslogs-search-limit'
 const STORAGE_VISIBLE_COLUMNS_KEY = 'dnslogs-visible-columns'
